feat: temperature conversion

+++ WebApp/Models/Device/StatusModel.cs
@@ -3,5 +3,5 @@ namespace Gas.WebApp.Models;
 public class StatusModel : BaseModel
 {
     public List<(DateTime, int)> Rssis { get; set; } = new();
-    public List<(DateTime, int)> SystemTemperatures { get; set; } = new();
+    public List<(DateTime, double)> SystemTemperatures { get; set; } = new();
 }
